requirements.txt
Flask==2.1.1
openai==0.27.0
Pillow==9.0.1
